ml/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
xgboost==2.1.*
shap==0.46.*
scikit-learn==1.6.*
pandas==2.2.*
psycopg2-binary==2.9.*
joblib==1.4.*
numpy==2.1.*
